internal/integration/prosody: run as root
It would be preferable (even in CI) to run Prosody as a normal user.
Unfortunately doing this in the integration testing code doesn't work
for reasons that are still unclear, we don't want to always be using
sudo, doas, or similar which may not exist on all systems, and we can't
use prosodyctl because it always daemonizes prosody.

For now, just allow Prosody to run as root.

Signed-off-by: Sam Whited <sam@samwhited.com>