.woodpecker: fix CI that was broken by Alpine
We were using a package out of the testing repository (py3-tzlocal, a
dependency of py3-aioxmpp) that was recently moved to the community
repository and is therefore only in the edge version of Alpine. Adding
the edge community repo can work around the problem until it makes its
way down into the version of alpine used by golang:alpine.

Signed-off-by: Sam Whited <sam@samwhited.com>