.woodpecker: use alpine:edge for integration tests
Previously we were using golang:alpine and adding the edge community and
testing repos to the apk configuration to ensure that the packages we
needed were available. Unfortunately this sometimes causes failures
because packages required by a package in community may not exist in
3.16's main branch, for example. We could add all three edge repos, but
it's safer to simply use edge and install Go ourselves.

Signed-off-by: Sam Whited <sam@samwhited.com>