internal/integration: add integration.skip flag
This patch adds a new flag that skips specific integration tests by
command name. It also disables the automatic skipping if the command
cannot be found (this will now cause tests to error).
This is a more explicit way to skip tests and means that it won't appear
as if CI is passing when we add a new command to test against but the
command to execute isn't installed.
The CI config has also been updated to use the new flag.

Signed-off-by: Sam Whited <sam@samwhited.com>