mux: ignore chardata in IQs
Handling of white space and character data in XML is confusing, and in
XMPP is underspecified. An issue was identified where some clients send
formatted XML, causing the multiplexer not to match on the start element
(because it would encounter a chardata with the value "\n" or similar
as the first token in the IQ instead of the start element). After a
discussion on jdev@muc.xmpp.org we determined that it was undefined and
that at least one other implementation (aioxmpp) was ignoring
non-significant whitespace, and indeed all chardata, inside of stanzas.
This patch updates the built-in multiplexer to do the same.

Fixes #315

Signed-off-by: Sam Whited <sam@samwhited.com>