internal/integration/jackal: randomize HTTP port
Previously the HTTP port was always the Jackal default, :6060. This
could lead to conflicts if we wanted to run multiple instances of Jackal
at once during the integration tests, or if one failed to shut down for
some reason. Instead, select a random HTTP port even though we're not
using any of Jackal's HTTP features currently.

Signed-off-by: Sam Whited <sam@samwhited.com>