xmpp: add UpdateAddr to set JID during bind
Previously the resource binding feature used internal session details,
making it impossible to implement outside of the main package. This is
the simplest API I could think of to allow updating the JID in a
relatively safe way that doesn't allow users to try and change it in the
middle of a session.

Fixes #104

Signed-off-by: Sam Whited <sam@samwhited.com>