xmpp: return better errors during StartTLS
In a very early design of the stream features API, errors returned from
the Negotiate function were marshaled and sent to the remote entity.
StartTLS still returns stream errors with very little info about what
was going on because of this, but also StartTLS should *never* transmit
an error (except the generic "failure" specifically defined for the
StartTLS feature).
Instead, return errors to the user that actually tells them what's going
on and immediately return (closing the connection).

Fixes #352

Signed-off-by: Sam Whited <sam@samwhited.com>