internal/ns: remove ns.WS constant
The ns.WS constant was a duplicate of the websocket namespace that was
created due to problems with the structure of websocket code that would
otherwise cause import loops.
Since it was only used in the stream package, and we intend to deprecate
the ns package in the future, remove the constant to reduce usage of ns.

Signed-off-by: Sam Whited <sam@samwhited.com>