stanza: use xmpptest for error encoding
This removes some duplicate test logic that existed before the xmpptest
encoding tests were a thing. Several tests that do not add any
additional coverage were left out, as well as two that led me to bugs in
the package. These will be added back in a future commit along with
their bug fix (in case they need to be rolled back we want separate
commits so that we don't have to roll back these test improvements with
them).

Signed-off-by: Sam Whited <sam@samwhited.com>