internal/genpubsub: update condition names
The last release accidentally released conditions without the special
"None" condition.
This meant that a breaking change was needed that would increment all
condition values by 1.
We also wanted a prefix on conditions to ensure they don't conflict
with anything, but weren't going to do that since it would break things.
Since we have to break things anyways, we might as well go all in.

Signed-off-by: Sam Whited <sam@samwhited.com>