muc: fix implicit memory aliasing
This fixes a bug where MUC history could wind up with an invalid value
for the "since" attribute that defines how long ago we're requesting
history for.
By copying the attribute value before we return a reference we are
ensuring that the previous value does not get overwritten in future
iterations of the loop and the correct attribute value is used.

Signed-off-by: Sam Whited <sam@samwhited.com>