stream: fix unmarshalling of stream error
Previously we were always skipping the end of child elements, even if we
had fully decoded that element and read to the end (as we did with
"text" elements, meaning we would skip over the next child after a
"text" element instead of just fast-forwarding to the end of the current
child as intended.

This improves CI to actually check that we get the correct value for the
text field as well.

Signed-off-by: Sam Whited <sam@samwhited.com>