.woodpecker: use "directory" pipeline syntax
Woodpecker CI provides us with a way to change the directory; previously
we were using cd but if we add other steps in the future these will be
in a fresh shell and we'd have to do that every time. Using the YAML
syntax means we'll always start in the examples directory for the
examples test.

Signed-off-by: Sam Whited <sam@samwhited.com>