.woodpecker: always build examples
Previously we were only building the examples when they changed.
However, now that the replace directive is back changing code in the
main library could break the examples. It only takes a few seconds to
build them all, so let's build it every time, regardless of whether they
have been changed or not.

Signed-off-by: Sam Whited <sam@samwhited.com>