dial: rework documentation
Previously the various options on the dialer had confusing or incorrect
documentation that could lead to incorrect use if not read carefully.
In addition, some identifiers had documentation that was confusing.or
was duplicated elsewhere that made more sense.

This patch cleans up the documentation and attempts to make it easier to
understand.

Signed-off-by: Sam Whited <sam@samwhited.com>