xmpp: add regression test for issue #399
There is currently a race condition during IQ handling that makes it
possible for the sender to time out and close the channel that's
listening for a response body after the stream handler has started
processing the response, causing it to attempt to write to the closed
channel. This test is a contrived way to force that behavior so that we
can verify any potential solutions.

Updates #399

Signed-off-by: Sam Whited <sam@samwhited.com>