crypto: fix broken tests
Apparently the SHA3 implementation is actually being linked in now, and
we were using that as an example of what would happen if you used a
crypto algorithm that was not actually in the binary. Update it to use
BLAKE2b_512 as an example since we're not likely to have any reason to
ever import that package.

Signed-off-by: Sam Whited <sam@samwhited.com>