ibb: fix deadlock in tests
TestExpect would wait on an incoming file transfer request, then send
the request and make sure that the Expect function unblocked.
Unfortunately, there was no easy way to guarantee that sending the
request happend *after* we were already waiting on the connection
request. The order was not guaranteed, so the test would occasionally
deadlock. This patch fixes the deadlock by adding a special method
that's only available in tests that blocks until the expected session ID
appears in the map of items being waited on.

Fixes #407

Signed-off-by: Sam Whited <sam@samwhited.com>