ibb: fix deadlock in tests
Like #409 where we fixed an error in some IBB tests due to the listener
starting concurrently with the actual request, this fixes a deadlock
where the listener was sometimes being created after a request was sent
to it.

This also updates the solution from #409 to use the obvious, simpler
solution that my sleepy brain didn't see this morning.

Updates #409
Fixes #410

Signed-off-by: Sam Whited <sam@samwhited.com>