xmpp: allow whitespace inside IQ stanza
We previously fixed the IQ stanza handling to allow IQ whitespace inside
the stanza, but missed the separate IQ handling during resource binding.
IIRC whitespace inside stanzas is technically illegal, but it appears
that some clients do it anyways so for compatibilities sake we may have
to deviate from the spec a bit here, at least for now.

Updates #288
Fixes #367

Signed-off-by: Sam Whited <sam@samwhited.com>