receipts: change API for requesting receipt
Before to request a read receipt we included a struct in our XML type
which would marshal to the request. This meant you had to have two types
if you wanted to toggle read receipts (one with the struct, one
without). The struct had a bool value on it so that it could be included
in types that would be unmarshaled and we could see if it was present.
This felt very inconsistent. Instead just make the entire type a bool
that can be set or unset at will to toggle read receipts.

Fixes #398

Signed-off-by: Sam Whited <sam@samwhited.com>