examples/commands: simplify ad-hoc commands UI
We had created our own label type for reasons that I don't remember
(maybe TextView didn't previously exist, or wasn't valid in forms), but
we can just use a TextView instead. This simplifies the UI so we can
focus on being an example of using ad-hoc commands and not an example of
using the UI toolkit the example uses.

Signed-off-by: Sam Whited <sam@samwhited.com>