dial: return a joined error if both lookups fail
When dialing an XMPP address we currently fetch SRV records for xmpp and
xmpps records concurrently. Previously if both lookups failed we were
returning one of the errors, even though in theory they could have
failed for different reasons and we'd want to know about both of them.
This patch uses errors.Join to return a new error that can be resolved
to each individual underlying error if necessary, but still ignores the
error and uses the successful lookup if only one of them failed.

Signed-off-by: Sam Whited <sam@samwhited.com>