dial, internal/discover: respect disabled fallback
The server may set a special DNS record to indicate that there is
explicitly no support for the service we're looking up at the given
domain. This patch respects that signal and does not attempt the
default fallback if this special record is seen, but does continue to
use the fallback if no record was set.

Signed-off-by: Sam Whited <sam@samwhited.com>