muc: fix race when joining channel
This patch fixes a race when joining a channel where the Go channel used
by the handler for signaling the completion of the join to the Join
method would be set to nil before the Join method had a chance to read
from it.

Instead, we now use a channel of channels and have the Join method
signal its own signaling channel to the handler. This has the bonus of
causing repeat joins to block and not send the Join stanza until the
previous re-join is finished.

Fixes #435

Signed-off-by: Sam Whited <sam@samwhited.com>